+++ nobodywho/core/src/tts/mod.rs
@@ -3,20 +3,23 @@
 /// The default public API accepts plain text and formats the required OuteTTS
 /// prompt internally. [`Tts::synthesize_prompt`] is available as an escape hatch
 /// for callers that need full control over the prompt.
+mod v02;
+mod v03;
+
 use crate::errors::{DecodingError, InitWorkerError, TtsWorkerError};
 use crate::llm::{self, Worker, WorkerGuard, GLOBAL_INFERENCE_LOCK};
 use crate::memory;
 use crate::sampler_config::SamplerPresets;
-use llama_cpp_2::context::LlamaContext;
-use llama_cpp_2::context::params::LlamaPoolingType;
 use llama_cpp_2::context::params::LlamaContextParams;
+use llama_cpp_2::context::params::LlamaPoolingType;
+use llama_cpp_2::context::LlamaContext;
 use llama_cpp_2::llama_batch::LlamaBatch;
 use llama_cpp_2::model::AddBos;
 use llama_cpp_2::token::LlamaToken;
 use serde::{Deserialize, Serialize};
 use stft_rs::{BatchIstft, ReconstructionMode, Spectrum, StftConfigBuilder, WindowType};
-use std::sync::Arc;
 use std::collections::VecDeque;
+use std::sync::Arc;
 use std::time::Instant;
 use tracing::{debug, error, info};
 
@@ -80,8 +83,8 @@ impl TtsBackendConfig {
 impl TextModelBackend {
     fn normalize_text(&self, text: &str) -> String {
         match self {
-            TextModelBackend::OuteTtsV02 => process_text_v02(text),
-            TextModelBackend::OuteTtsV03 => process_text_v03(text),
+            TextModelBackend::OuteTtsV02 => v02::process_text(text),
+            TextModelBackend::OuteTtsV03 => v03::process_text(text),
         }
     }
 
@@ -100,29 +103,16 @@ impl TextModelBackend {
     }
 
     fn cached_prompt_prefix(&self, speaker: &TtsSpeaker) -> Option<String> {
-        match (self, speaker) {
-            (TextModelBackend::OuteTtsV02, TtsSpeaker::Preset(TtsSpeakerPreset::DefaultEnglishMale)) => {
-                Some(default_english_male_prefix())
-            }
-            (TextModelBackend::OuteTtsV03, _) => None,
-            (TextModelBackend::OuteTtsV02, TtsSpeaker::Profile(_)) => None,
+        match self {
+            TextModelBackend::OuteTtsV02 => v02::cached_prompt_prefix(speaker),
+            TextModelBackend::OuteTtsV03 => None,
         }
     }
 
     fn prompt(&self, request: &PreparedTtsRequest) -> Result<String, TtsWorkerError> {
-        match (self, &request.speaker) {
-            (TextModelBackend::OuteTtsV02, TtsSpeaker::Preset(TtsSpeakerPreset::DefaultEnglishMale)) => {
-                Ok(build_default_english_male_prompt(&request.processed_text))
-            }
-            (TextModelBackend::OuteTtsV02, TtsSpeaker::Profile(_)) => Err(TtsWorkerError::InvalidRequest(
-                "speaker profiles are not supported by the OuteTTS 0.2 backend".into(),
-            )),
-            (TextModelBackend::OuteTtsV03, TtsSpeaker::Preset(_)) => {
-                Ok(build_outetts_v03_prompt_no_speaker(&request.processed_text))
-            }
-            (TextModelBackend::OuteTtsV03, TtsSpeaker::Profile(profile)) => {
-                Ok(build_outetts_v03_prompt(&request.processed_text, profile)?)
-            }
+        match self {
+            TextModelBackend::OuteTtsV02 => v02::prompt(request),
+            TextModelBackend::OuteTtsV03 => v03::prompt(request),
         }
     }
 
@@ -156,7 +146,7 @@ impl VocoderBackend {
 }
 
 #[derive(Clone, Debug)]
-struct PreparedTtsRequest {
+pub(crate) struct PreparedTtsRequest {
     processed_text: String,
     spoken_char_count: usize,
     speaker: TtsSpeaker,
@@ -214,7 +204,7 @@ impl TtsSpeakerProfile {
         Self::from_json_str(&json)
     }
 
-    fn validate(&self) -> Result<(), TtsWorkerError> {
+    pub(crate) fn validate(&self) -> Result<(), TtsWorkerError> {
         if self.words.is_empty() {
             return Err(TtsWorkerError::InvalidRequest(
                 "speaker profile must contain at least one word".into(),
@@ -331,17 +321,14 @@ impl Tts {
         })
     }
 
-    /// Synthesize speech from plain text using the default prompt builder.
     pub fn synthesize(&self, text: impl Into<String>) -> Result<Vec<u8>, TtsWorkerError> {
         futures::executor::block_on(self.async_handle.synthesize(text))
     }
 
-    /// Synthesize speech from a structured text request.
     pub fn synthesize_request(&self, request: TtsRequest) -> Result<Vec<u8>, TtsWorkerError> {
         futures::executor::block_on(self.async_handle.synthesize_request(request))
     }
 
-    /// Synthesize speech from a fully preformatted model prompt.
     pub fn synthesize_prompt(&self, prompt: impl Into<String>) -> Result<Vec<u8>, TtsWorkerError> {
         futures::executor::block_on(self.async_handle.synthesize_prompt(prompt))
     }
@@ -432,12 +419,10 @@ impl TtsAsync {
         }
     }
 
-    /// Synthesize speech from plain text using the default prompt builder.
     pub async fn synthesize(&self, text: impl Into<String>) -> Result<Vec<u8>, TtsWorkerError> {
         self.synthesize_request(TtsRequest::new(text)).await
     }
 
-    /// Synthesize speech from a structured text request.
     pub async fn synthesize_request(&self, request: TtsRequest) -> Result<Vec<u8>, TtsWorkerError> {
         let (result_tx, mut result_rx) = tokio::sync::mpsc::channel(1);
         self.guard
@@ -448,7 +433,6 @@ impl TtsAsync {
         result_rx.recv().await.ok_or(TtsWorkerError::NoResponse)?
     }
 
-    /// Synthesize speech from a fully preformatted model prompt.
     pub async fn synthesize_prompt(
         &self,
         prompt: impl Into<String>,
@@ -524,9 +508,6 @@ impl<'a> Worker<'a, VocoderWorker> {
         n_ctx: u32,
         backend: VocoderBackend,
     ) -> Result<Worker<'a, VocoderWorker>, InitWorkerError> {
-        // The vocoder encodes the full audio-code sequence in one pass and then
-        // reads per-token embeddings back out, so n_ubatch must cover the active
-        // sequence length for longer utterances.
         let ctx = new_vocoder_context(model, n_ctx, &backend)?;
         let big_batch = LlamaBatch::new(ctx.n_ctx() as usize, 1);
         let small_batch = LlamaBatch::new(1, 1);
@@ -711,7 +692,7 @@ fn synthesize_prepared_inner(
     match &request.speaker {
         TtsSpeaker::Preset(TtsSpeakerPreset::DefaultEnglishMale) => {
             if let Some(cache) = tts_worker.extra.default_english_male.clone() {
-                let suffix = build_default_english_male_suffix(&request.processed_text);
+                let suffix = v02::build_default_english_male_suffix(&request.processed_text);
                 return synthesize_with_cached_prefix(
                     tts_worker,
                     audio_decoder,
@@ -798,7 +779,6 @@ fn synthesize_with_cached_prefix(
     Ok(wav)
 }
 
-/// Run the OuteTTS language model autoregressively, collecting audio code tokens.
 fn generate_audio_codes(
     worker: &mut Worker<'_, TtsWorker>,
     prompt: String,
@@ -835,7 +815,6 @@ fn generate_audio_codes_from_current_state(
         ));
     }
 
-    // Hold the global inference lock for the entire sampling loop.
     let _lock = GLOBAL_INFERENCE_LOCK.lock().unwrap();
 
     let mut sampler = worker
@@ -909,7 +888,6 @@ fn generate_audio_codes_from_current_state(
             debug!(terminal_reason = stop_reason, "terminal token reached, stopping generation");
             break;
         }
-
     }
 
     let recent_tokens = recent_tokens.into_iter().collect::<Vec<_>>().join(" ");
@@ -921,6 +899,24 @@ fn generate_audio_codes_from_current_state(
     );
 
     Ok(audio_codes)
+}
+
+fn resolve_special_token(
+    worker: &Worker<'_, TtsWorker>,
+    piece: &str,
+) -> Result<LlamaToken, TtsWorkerError> {
+    let tokens = worker
+        .ctx
+        .model
+        .str_to_token(piece, AddBos::Never)
+        .map_err(|e| TtsWorkerError::InvalidRequest(format!("failed to tokenize {piece}: {e}")))?;
+
+    match tokens.as_slice() {
+        [token] => Ok(*token),
+        _ => Err(TtsWorkerError::InvalidRequest(format!(
+            "{piece} did not resolve to a single token"
+        ))),
+    }
 }
 
 fn push_recent_token(
@@ -946,24 +942,6 @@ fn debug_token_piece(worker: &Worker<'_, TtsWorker>, token: LlamaToken) -> Strin
     match token_bytes {
         Ok(bytes) => String::from_utf8_lossy(&bytes).replace('\n', "\\n"),
         Err(_) => "<detok_err>".into(),
-    }
-}
-
-fn resolve_special_token(
-    worker: &Worker<'_, TtsWorker>,
-    piece: &str,
-) -> Result<LlamaToken, TtsWorkerError> {
-    let tokens = worker
-        .ctx
-        .model
-        .str_to_token(piece, AddBos::Never)
-        .map_err(|e| TtsWorkerError::InvalidRequest(format!("failed to tokenize {piece}: {e}")))?;
-
-    match tokens.as_slice() {
-        [token] => Ok(*token),
-        _ => Err(TtsWorkerError::InvalidRequest(format!(
-            "{piece} did not resolve to a single token"
-        ))),
     }
 }
 
@@ -998,7 +976,6 @@ fn initialize_tts_caches(worker: &mut Worker<'_, TtsWorker>) -> Result<(), TtsWo
     Ok(())
 }
 
-/// Run the WavTokenizer vocoder and return per-token spectral embeddings.
 fn run_vocoder(
     worker: &mut Worker<'_, VocoderWorker>,
     audio_codes: &[LlamaToken],
@@ -1026,10 +1003,6 @@ fn run_vocoder(
         .encode(&mut worker.big_batch)
         .map_err(|e| TtsWorkerError::Vocoder(e.to_string()))?;
 
-    // llama-cpp-2's embeddings_ith() uses model.n_embd() (=features_length=512) to size the
-    // returned slice, but for WavTokenizer the actual per-token output is n_embd_out
-    // (=embedding_length=1282). Read the correct size from GGUF metadata and use the raw
-    // llama_get_embeddings_ith pointer directly with the right length.
     let n_embd_out = worker
         .ctx
         .model
@@ -1049,9 +1022,6 @@ fn run_vocoder(
             .ctx
             .embeddings_ith(i as i32)
             .map_err(|e| TtsWorkerError::Vocoder(e.to_string()))?;
-        // llama-cpp-2's embeddings_ith() sizes the slice to model.n_embd() (=features_length=512),
-        // but for WavTokenizer the actual per-token output is n_embd_out (=embedding_length=1282).
-        // We extend the slice to read all n_embd_out values from the same underlying buffer.
         let extended = unsafe { std::slice::from_raw_parts(emb.as_ptr(), n_embd_out) };
         frames.push(extended.to_vec());
     }
@@ -1059,10 +1029,6 @@ fn run_vocoder(
     Ok(frames)
 }
 
-/// Reconstruct a PCM waveform from WavTokenizer spectral frames using stft-rs WOLA.
-///
-/// Each frame is a flat vector of [log_magnitudes | phases] (n_embd/2 values each).
-/// WOLA with a Hann window matches the energy-normalized OLA used by tts.cpp.
 #[cfg(test)]
 fn reconstruct_audio(frames: &[Vec<f32>]) -> Result<Vec<f32>, TtsWorkerError> {
     reconstruct_audio_with_backend(frames, &VocoderBackend::WavTokenizer75)
@@ -1087,8 +1053,8 @@ fn reconstruct_audio_with_backend(
             "inconsistent vocoder embedding widths across frames".into(),
         ));
     }
-    let n_bins = n_embd / 2; // 641 for n_embd=1282
-    let n_fft = (n_bins - 1) * 2; // 1280
+    let n_bins = n_embd / 2;
+    let n_fft = (n_bins - 1) * 2;
 
     let config = StftConfigBuilder::<f32>::new()
         .fft_size(n_fft)
@@ -1101,15 +1067,9 @@ fn reconstruct_audio_with_backend(
     let istft = BatchIstft::new(config);
     let n_codes = frames.len();
 
-    // The WOLA ISTFT only outputs n_codes * hop_size samples, but the last frame's
-    // window extends (n_fft - hop_size) samples beyond that, cutting off the audio
-    // tail. Append silent flush frames so the reconstruction emits the full tail.
-    let n_flush = n_fft / backend.hop_size(); // 4 for WavTokenizer75 (1280/320)
+    let n_flush = n_fft / backend.hop_size();
     let n_frames = n_codes + n_flush;
 
-    // stft-rs layout: data[0..n_frames*n_bins] = all real parts (row-major),
-    //                 data[n_frames*n_bins..]  = all imaginary parts.
-    // Flush frames remain zero (already zero-initialised).
     let mut data = vec![0.0f32; 2 * n_frames * n_bins];
     for (i, frame) in frames.iter().enumerate() {
         let log_mags = &frame[..n_bins];
@@ -1157,204 +1117,6 @@ fn encode_wav(pcm: &[f32]) -> Result<Vec<u8>, TtsWorkerError> {
     Ok(buffer)
 }
 
-fn default_english_male_prefix() -> String {
-    let reference_text = "the<|text_sep|>overall<|text_sep|>package<|text_sep|>from<|text_sep|>\
-        just<|text_sep|>two<|text_sep|>people<|text_sep|>is<|text_sep|>pretty<|text_sep|>\
-        remarkable<|text_sep|>sure<|text_sep|>i<|text_sep|>have<|text_sep|>some<|text_sep|>\
-        critiques<|text_sep|>about<|text_sep|>some<|text_sep|>of<|text_sep|>the<|text_sep|>\
-        gameplay<|text_sep|>aspects<|text_sep|>but<|text_sep|>its<|text_sep|>still<|text_sep|>\
-        really<|text_sep|>enjoyable<|text_sep|>and<|text_sep|>it<|text_sep|>looks<|text_sep|>\
-        lovely";
-
-    let reference_audio = r#"<|audio_start|>
-the<|t_0.08|><|code_start|><|257|><|740|><|636|><|913|><|788|><|1703|><|code_end|>
-overall<|t_0.36|><|code_start|><|127|><|201|><|191|><|774|><|700|><|532|><|1056|><|557|><|798|><|298|><|1741|><|747|><|1662|><|1617|><|1702|><|1527|><|368|><|1588|><|1049|><|1008|><|1625|><|747|><|1576|><|728|><|1019|><|1696|><|1765|><|code_end|>
-package<|t_0.56|><|code_start|><|935|><|584|><|1319|><|627|><|1016|><|1491|><|1344|><|1117|><|1526|><|1040|><|239|><|1435|><|951|><|498|><|723|><|1180|><|535|><|789|><|1649|><|1637|><|78|><|465|><|1668|><|901|><|595|><|1675|><|117|><|1009|><|1667|><|320|><|840|><|79|><|507|><|1762|><|1508|><|1228|><|1768|><|802|><|1450|><|1457|><|232|><|639|><|code_end|>
-from<|t_0.19|><|code_start|><|604|><|782|><|1682|><|872|><|1532|><|1600|><|1036|><|1761|><|647|><|1554|><|1371|><|653|><|1595|><|950|><|code_end|>
-just<|t_0.25|><|code_start|><|1782|><|1670|><|317|><|786|><|1748|><|631|><|599|><|1155|><|1364|><|1524|><|36|><|1591|><|889|><|1535|><|541|><|440|><|1532|><|50|><|870|><|code_end|>
-two<|t_0.24|><|code_start|><|1681|><|1510|><|673|><|799|><|805|><|1342|><|330|><|519|><|62|><|640|><|1138|><|565|><|1552|><|1497|><|1552|><|572|><|1715|><|1732|><|code_end|>
-people<|t_0.39|><|code_start|><|593|><|274|><|136|><|740|><|691|><|633|><|1484|><|1061|><|1138|><|1485|><|344|><|428|><|397|><|1562|><|645|><|917|><|1035|><|1449|><|1669|><|487|><|442|><|1484|><|1329|><|1832|><|1704|><|600|><|761|><|653|><|269|><|code_end|>
-is<|t_0.16|><|code_start|><|566|><|583|><|1755|><|646|><|1337|><|709|><|802|><|1008|><|485|><|1583|><|652|><|10|><|code_end|>
-pretty<|t_0.32|><|code_start|><|1818|><|1747|><|692|><|733|><|1010|><|534|><|406|><|1697|><|1053|><|1521|><|1355|><|1274|><|816|><|1398|><|211|><|1218|><|817|><|1472|><|1703|><|686|><|13|><|822|><|445|><|1068|><|code_end|>
-remarkable<|t_0.68|><|code_start|><|230|><|1048|><|1705|><|355|><|706|><|1149|><|1535|><|1787|><|1356|><|1396|><|835|><|1583|><|486|><|1249|><|286|><|937|><|1076|><|1150|><|614|><|42|><|1058|><|705|><|681|><|798|><|934|><|490|><|514|><|1399|><|572|><|1446|><|1703|><|1346|><|1040|><|1426|><|1304|><|664|><|171|><|1530|><|625|><|64|><|1708|><|1830|><|1030|><|443|><|1509|><|1063|><|1605|><|1785|><|721|><|1440|><|923|><|code_end|>
-sure<|t_0.36|><|code_start|><|792|><|1780|><|923|><|1640|><|265|><|261|><|1525|><|567|><|1491|><|1250|><|1730|><|362|><|919|><|1766|><|543|><|1|><|333|><|113|><|970|><|252|><|1606|><|133|><|302|><|1810|><|1046|><|1190|><|1675|><|code_end|>
-i<|t_0.08|><|code_start|><|123|><|439|><|1074|><|705|><|1799|><|637|><|code_end|>
-have<|t_0.16|><|code_start|><|1509|><|599|><|518|><|1170|><|552|><|1029|><|1267|><|864|><|419|><|143|><|1061|><|0|><|code_end|>
-some<|t_0.16|><|code_start|><|619|><|400|><|1270|><|62|><|1370|><|1832|><|917|><|1661|><|167|><|269|><|1366|><|1508|><|code_end|>
-critiques<|t_0.60|><|code_start|><|559|><|584|><|1163|><|1129|><|1313|><|1728|><|721|><|1146|><|1093|><|577|><|928|><|27|><|630|><|1080|><|1346|><|1337|><|320|><|1382|><|1175|><|1682|><|1556|><|990|><|1683|><|860|><|1721|><|110|><|786|><|376|><|1085|><|756|><|1523|><|234|><|1334|><|1506|><|1578|><|659|><|612|><|1108|><|1466|><|1647|><|308|><|1470|><|746|><|556|><|1061|><|code_end|>
-about<|t_0.29|><|code_start|><|26|><|1649|><|545|><|1367|><|1263|><|1728|><|450|><|859|><|1434|><|497|><|1220|><|1285|><|179|><|755|><|1154|><|779|><|179|><|1229|><|1213|><|922|><|1774|><|1408|><|code_end|>
-some<|t_0.23|><|code_start|><|986|><|28|><|1649|><|778|><|858|><|1519|><|1|><|18|><|26|><|1042|><|1174|><|1309|><|1499|><|1712|><|1692|><|1516|><|1574|><|code_end|>
-of<|t_0.07|><|code_start|><|197|><|716|><|1039|><|1662|><|64|><|code_end|>
-the<|t_0.08|><|code_start|><|1811|><|1568|><|569|><|886|><|1025|><|1374|><|code_end|>
-gameplay<|t_0.48|><|code_start|><|1269|><|1092|><|933|><|1362|><|1762|><|1700|><|1675|><|215|><|781|><|1086|><|461|><|838|><|1022|><|759|><|649|><|1416|><|1004|><|551|><|909|><|787|><|343|><|830|><|1391|><|1040|><|1622|><|1779|><|1360|><|1231|><|1187|><|1317|><|76|><|997|><|989|><|978|><|737|><|189|><|code_end|>
-aspects<|t_0.56|><|code_start|><|1423|><|797|><|1316|><|1222|><|147|><|719|><|1347|><|386|><|1390|><|1558|><|154|><|440|><|634|><|592|><|1097|><|1718|><|712|><|763|><|1118|><|1721|><|1311|><|868|><|580|><|362|><|1435|><|868|><|247|><|221|><|886|><|1145|><|1274|><|1284|><|457|><|1043|><|1459|><|1818|><|62|><|599|><|1035|><|62|><|1649|><|778|><|code_end|>
-but<|t_0.20|><|code_start|><|780|><|1825|><|1681|><|1007|><|861|><|710|><|702|><|939|><|1669|><|1491|><|613|><|1739|><|823|><|1469|><|648|><|code_end|>
-its<|t_0.09|><|code_start|><|92|><|688|><|1623|><|962|><|1670|><|527|><|599|><|code_end|>
-still<|t_0.27|><|code_start|><|636|><|10|><|1217|><|344|><|713|><|957|><|823|><|154|><|1649|><|1286|><|508|><|214|><|1760|><|1250|><|456|><|1352|><|1368|><|921|><|615|><|5|><|code_end|>
-really<|t_0.36|><|code_start|><|55|><|420|><|1008|><|1659|><|27|><|644|><|1266|><|617|><|761|><|1712|><|109|><|1465|><|1587|><|503|><|1541|><|619|><|197|><|1019|><|817|><|269|><|377|><|362|><|1381|><|507|><|1488|><|4|><|1695|><|code_end|>
-enjoyable<|t_0.49|><|code_start|><|678|><|501|><|864|><|319|><|288|><|1472|><|1341|><|686|><|562|><|1463|><|619|><|1563|><|471|><|911|><|730|><|1811|><|1006|><|520|><|861|><|1274|><|125|><|1431|><|638|><|621|><|153|><|876|><|1770|><|437|><|987|><|1653|><|1109|><|898|><|1285|><|80|><|593|><|1709|><|843|><|code_end|>
-and<|t_0.15|><|code_start|><|1285|><|987|><|303|><|1037|><|730|><|1164|><|502|><|120|><|1737|><|1655|><|1318|><|code_end|>
-it<|t_0.09|><|code_start|><|848|><|1366|><|395|><|1601|><|1513|><|593|><|1302|><|code_end|>
-looks<|t_0.27|><|code_start|><|1281|><|1266|><|1755|><|572|><|248|><|1751|><|1257|><|695|><|1380|><|457|><|659|><|585|><|1315|><|1105|><|1776|><|736|><|24|><|736|><|654|><|1027|><|code_end|>
-lovely<|t_0.56|><|code_start|><|634|><|596|><|1766|><|1556|><|1306|><|1285|><|1481|><|1721|><|1123|><|438|><|1246|><|1251|><|795|><|659|><|1381|><|1658|><|217|><|1772|><|562|><|952|><|107|><|1129|><|1112|><|467|><|550|><|1079|><|840|><|1615|><|1469|><|1380|><|168|><|917|><|836|><|1827|><|437|><|583|><|67|><|595|><|1087|><|1646|><|1493|><|1677|><|code_end|>
-<|audio_end|>
-"#;
-
-    format!(
-        "<|im_start|>\n<|text_start|>{reference_text}<|text_end|>\n{reference_audio}"
-    )
-}
-
-fn build_default_english_male_suffix(processed_text: &str) -> String {
-    format!("<|text_start|>{processed_text}<|text_end|>\n<|audio_start|>\n")
-}
-
-fn build_default_english_male_prompt(processed_text: &str) -> String {
-    format!(
-        "{}{}",
-        default_english_male_prefix(),
-        build_default_english_male_suffix(processed_text)
-    )
-}
-
-fn build_outetts_v03_prompt_no_speaker(processed_text: &str) -> String {
-    format!("<|im_start|>\n<|text_start|>{processed_text}<|text_end|>\n<|audio_start|>\n")
-}
-
-fn build_outetts_v03_prompt(
-    processed_text: &str,
-    profile: &TtsSpeakerProfile,
-) -> Result<String, TtsWorkerError> {
-    profile.validate()?;
-
-    // Text section: speaker words + space + target words
-    let speaker_text: String = profile
-        .words
-        .iter()
-        .map(|w| process_text_v03(&w.word))
-        .collect::<Vec<_>>()
-        .join("<|space|>");
-
-    let full_text = format!("{}<|space|>{}", speaker_text, processed_text);
-
-    // Audio section: speaker audio codes as prefix (model will continue with target)
-    let mut audio_prefix = String::new();
-    for (i, word) in profile.words.iter().enumerate() {
-        if i > 0 {
-            audio_prefix.push_str("<|space|>\n");
-        }
-        audio_prefix.push_str(&serialize_outetts_v03_profile_word(word)?);
-    }
-    audio_prefix.push_str("<|space|>\n");
-
-    Ok(format!(
-        "<|im_start|>\n<|text_start|>{full_text}<|text_end|>\n<|audio_start|>\n{audio_prefix}"
-    ))
-}
-
-fn serialize_outetts_v03_profile_word(word: &TtsSpeakerWord) -> Result<String, TtsWorkerError> {
-    let normalized_word = process_text_v03(&word.word);
-    if normalized_word.is_empty() {
-        return Err(TtsWorkerError::InvalidRequest(format!(
-            "speaker profile word {:?} becomes empty after normalization",
-            word.word
-        )));
-    }
-
-    let mut serialized = normalized_word;
-    serialized.push_str(&format_duration_token(word.duration));
-    for &code in &word.codes {
-        serialized.push_str(&format!("<|{code}|>"));
-    }
-    Ok(serialized)
-}
-
-fn format_duration_token(duration: f32) -> String {
-    let centiseconds = (duration.max(0.0) * 100.0).round() as i32;
-    let centiseconds = centiseconds.clamp(0, 1000);
-    format!("<|t_{}.{:02}|>", centiseconds / 100, centiseconds % 100)
-}
-
-fn process_text_v02(text: &str) -> String {
-    let s = text.to_lowercase();
-
-    let s: String = s
-        .chars()
-        .map(|c| match c {
-            '-' | '_' | '/' | ',' | '.' | '\\' => ' ',
-            c => c,
-        })
-        .collect();
-
-    let s: String = s
-        .chars()
-        .filter(|c| c.is_ascii_alphabetic() || *c == ' ')
-        .collect();
-
-    let s = s.split_whitespace().collect::<Vec<_>>().join(" ");
-    s.replace(' ', "<|text_sep|>")
-}
-
-fn process_text_v03(text: &str) -> String {
-    let mut normalized = String::new();
-    let mut pending_space = false;
-    let chars: Vec<char> = text.to_lowercase().chars().collect();
-    let mut i = 0;
-
-    while i < chars.len() {
-        let c = chars[i];
-
-        if i + 2 < chars.len() && c == '.' && chars[i + 1] == '.' && chars[i + 2] == '.' {
-            normalized.push_str("<|ellipsis|>");
-            pending_space = false;
-            i += 3;
-            continue;
-        }
-
-        let mapped = match c {
-            '.' => Some("<|period|>"),
-            '!' => Some("<|exclamation_mark|>"),
-            '?' => Some("<|question_mark|>"),
-            ',' => Some("<|comma|>"),
-            '"' => Some("<|double_quote|>"),
-            '„' => Some("<|low_double_quote|>"),
-            '¡' => Some("<|inverted_exclamation|>"),
-            '¿' => Some("<|inverted_question|>"),
-            '…' => Some("<|ellipsis|>"),
-            '。' => Some("<|cjk_period|>"),
-            '！' => Some("<|cjk_exclamation|>"),
-            '？' => Some("<|cjk_question|>"),
-            '，' => Some("<|cjk_comma|>"),
-            '؟' => Some("<|arabic_question|>"),
-            _ => None,
-        };
-
-        if let Some(token) = mapped {
-            normalized.push_str(token);
-            pending_space = false;
-            i += 1;
-            continue;
-        }
-
-        match c {
-            '-' | '_' | '/' | '\\' => pending_space = !normalized.is_empty(),
-            c if c.is_whitespace() => pending_space = !normalized.is_empty(),
-            c if c.is_alphabetic() => {
-                if pending_space && !normalized.is_empty() {
-                    normalized.push(' ');
-                }
-                normalized.push(c);
-                pending_space = false;
-            }
-            _ => {}
-        }
-
-        i += 1;
-    }
-
-    normalized.replace(' ', "<|space|>")
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -1370,13 +1132,13 @@ mod tests {
 
     #[test]
     fn test_process_text_v02() {
-        let processed = process_text_v02("Hello, Rust-world 123!");
+        let processed = v02::process_text("Hello, Rust-world 123!");
         assert_eq!(processed, "hello<|text_sep|>rust<|text_sep|>world");
     }
 
     #[test]
     fn test_process_text_v03() {
-        let processed = process_text_v03("Hello, world! Really...");
+        let processed = v03::process_text("Hello, world! Really...");
         assert_eq!(
             processed,
             "hello<|comma|><|space|>world<|exclamation_mark|><|space|>really<|ellipsis|>"
@@ -1387,7 +1149,7 @@ mod tests {
     fn test_request_builds_prompt() -> Result<(), Box<dyn std::error::Error>> {
         let prepared = TtsRequest::new("Hello world")
             .prepare_with_backend(&TextModelBackend::OuteTtsV02)?;
-        let prompt = build_default_english_male_prompt(&prepared.processed_text);
+        let prompt = v02::prompt(&prepared)?;
         assert!(prompt.contains("<|audio_start|>"));
         assert!(prompt.contains("<|text_start|>hello<|text_sep|>world<|text_end|>"));
         Ok(())
@@ -1395,7 +1157,6 @@ mod tests {
 
     #[test]
     fn test_outetts_v03_no_speaker_builds_prompt() -> Result<(), Box<dyn std::error::Error>> {
-        // v0.3 without a speaker profile should produce a valid prompt (default voice)
         let prepared = TtsRequest::new("Hello world")
             .prepare_with_backend(&TextModelBackend::OuteTtsV03)?;
         let prompt = TextModelBackend::OuteTtsV03.prompt(&prepared)?;
@@ -1426,13 +1187,10 @@ mod tests {
             .with_speaker_profile(profile)
             .prepare_with_backend(&TextModelBackend::OuteTtsV03)?;
         let prompt = TextModelBackend::OuteTtsV03.prompt(&prepared)?;
-        // Speaker words + target words appear together in the text section
         assert!(prompt.contains("<|text_start|>hello<|space|>world<|space|>hello<|space|>world<|text_end|>"));
-        // Speaker audio codes appear in the audio section as a prefix
         assert!(prompt.contains("<|audio_start|>"));
         assert!(prompt.contains("hello<|t_0.52|><|551|><|552|>"));
         assert!(prompt.contains("world<|t_0.25|><|3|>"));
-        // The old voice_characteristic_start format is no longer used
         assert!(!prompt.contains("<|voice_characteristic_start|>"));
         Ok(())
     }
